DAOS-16304 control: Move metric types into lib/daos
Allow these types to be used without importing all of the Control
API.

Required-githooks: true
